move AvlNode defn into map.c
