--- src/node.h
@@ -1,48 +1,0 @@
-/*  MIT License
- *
- *  Copyright (c) 2019 Gregory Meyer
- *
- *  Permission is hereby granted, free of charge, to any person
- *  obtaining a copy of this software and associated documentation
- *  files (the "Software"), to deal in the Software without
- *  restriction, including without limitation the rights to use, copy,
- *  modify, merge, publish, distribute, sublicense, and/or sell copies
- *  of the Software, and to permit persons to whom the Software is
- *  furnished to do so, subject to the following conditions:
- *
- *  The above copyright notice and this permission notice (including
- *  the next paragraph) shall be included in all copies or substantial
- *  portions of the Software.
- *
- *  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
- *  EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
- *  MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
- *  NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS
- *  BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN
- *  ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
- *  CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
- *  IN THE SOFTWARE.
- */
-
-#ifndef AVL_INTERNAL_NODE_H
-#define AVL_INTERNAL_NODE_H
-
-#include <avlbst.h>
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-struct AvlNode {
-    AvlNode *left;
-    AvlNode *right;
-    void *key;
-    void *value;
-    signed char balance_factor;
-};
-
-#ifdef __cplusplus
-} // extern "C"
-#endif
-
-#endif
